+++ src/material/standard/standard-material.ts
@@ -126,19 +126,6 @@ export class StandardMaterial extends Material {
   set alphaMode(value: StandardMaterialAlphaMode) {
     if (this._alphaMode !== value) {
       this._alphaMode = value
-      switch (this._alphaMode) {
-        case StandardMaterialAlphaMode.opaque:
-        case StandardMaterialAlphaMode.mask: {
-          this.renderSortType = MaterialRenderSortType.opaque
-          this.state.depthMask = true; break
-        }
-        case StandardMaterialAlphaMode.blend: {
-          this.renderSortType = MaterialRenderSortType.transparent
-          // Depth mask feature is only available in PixiJS 6.0+ and won't have
-          // any effects in previous versions.
-          this.state.depthMask = false; break
-        }
-      }
       this.invalidateShader()
     }
   }
@@ -246,6 +233,7 @@ export class StandardMaterial extends Material {
       switch (source.alphaMode) {
         case "BLEND": {
           material.alphaMode = StandardMaterialAlphaMode.blend
+          material.renderSortType = MaterialRenderSortType.transparent
           break
         }
         case "MASK": {